implements JWT authentication

+++ src/main/kotlin/com/fourall/guildapi/repository/EventRepository.kt
@@ -3,6 +3,6 @@ package com.fourall.guildapi.repository
 import com.fourall.guildapi.model.Event
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface EventRepository : JpaRepository<Event, Int> {
+interface EventRepository : JpaRepository<Event, Long> {
     fun findByName(name: String) : Event
 }
